refactor(journal): rename exception to reflect package
