+++ journal/src/main/java/io/camunda/zeebe/journal/CorruptedJournalException.java
@@ -5,21 +5,25 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.journal.record;
+package io.camunda.zeebe.journal;
 
 import io.camunda.zeebe.util.exception.UnrecoverableException;
 
-public final class CorruptedLogException extends UnrecoverableException {
+/**
+ * Represents a failure to read or write to the journal which resulted in an inconsistency. This is
+ * an unrecoverable error which most likely requires manual intervention to fix (if possible).
+ */
+public final class CorruptedJournalException extends UnrecoverableException {
 
-  public CorruptedLogException(final String message) {
+  public CorruptedJournalException(final String message) {
     super(message);
   }
 
-  public CorruptedLogException(final Throwable cause) {
+  public CorruptedJournalException(final Throwable cause) {
     super(cause);
   }
 
-  public CorruptedLogException(final String message, final Throwable cause) {
+  public CorruptedJournalException(final String message, final Throwable cause) {
     super(message, cause);
   }
 }
